Fix handling of --indir in thraxcompiler.
GrmCompilerSpec<>::ParseFile() is used to process the main grammar file
as well as imports. The paths of imports will have already been
prepended with FLAGS_indir. As a result, the --indir path was duplicated
when searching for files being imported. Put differently, imports
only worked properly when --indir was set to the empty string. The present
change fixes this.